refactor(search): extract reindex business logic to a service (route 418→52)
The reindex-embeddings API route was 418 LOC (2.8× the 150 limit) doing all the
business logic inline — a clear SoC violation (audit finding). Moved it to
src/services/search/reindexService.ts:
- reconcileOne (single-row, DB-trigger path)
- reconcileCorpus (incremental sweep / full re-embed) + buildCorpus helper
- quality scoring, ENTITY_CFG (registry-derived), batching

The route is now a thin auth(x-reindex-secret) + embeddings-guard + delegate
wrapper (52 LOC). Behavior-preserving. Bonus: the corpus entity loop now derives
tables/paths from ENTITY_CFG (registry SSOT) instead of the hardcoded
user_products/user_services/user_causes + basePath literals it had.

tsc 0 errors; schema-drift gate green; no raw table literals.
